Clearing up main and putting non-core functions into utitilty file
