+++ main.py
@@ -1,52 +1,14 @@
-import cv2
 import numpy as np
 import pyautogui
-import autopy
 import time
 import threading
 import sys
 import logging
+from utility import *
 
+# Icons don't work across device so change this as needed
 device = 'laptop'
 #device = 'pc'
-
-# Quick click function, more reliable than pyautogui.click()
-def click(x, y):
-    pyautogui.moveTo(x, y)
-    pyautogui.mouseDown()
-    pyautogui.mouseUp()
-
-def locate_function(image, side, conf):
-    if side:
-        # When side is True limit the region to left side of screen, this is where user sprites are
-        x, y = pyautogui.locateCenterOnScreen(image, confidence=conf, region=(320, 0, 700, 1000))
-    else:
-        # By default use whole screen
-        x, y = pyautogui.locateCenterOnScreen(image, confidence=conf)
-    return x, y
-
-def find_and_click(image, user_side = False, x_inc=1, y_inc=10, x_mult=1, y_mult=1, confidence=0.8):
-    # Name of image on screen to find
-    image_to_find = 'images/'+device+'/'+image+'.png'
-    # Get coordinates of center point of image on screen
-    x_raw, y_raw = locate_function(image_to_find, user_side, confidence)
-
-    # Offset coordinates a little for hitbox
-    x_offset = x_raw + x_inc * x_mult
-    y_offset = y_raw + y_inc * y_mult
-    click(x_offset, y_offset)
-
-# Used to take screenshots for debugging
-def screenies():
-    i = 0
-    t = time.time()
-    while True:
-        myScreenshot = pyautogui.screenshot()
-        myScreenshot.save(f'images/'+device+'/screenshots/image{i}.png')
-        i+=1
-        time.sleep(.01) # Wait 10ms
-        if time.time()-t>=3: # Take screenshots for 3 seconds total
-            break
 
 def run_game():
     global game_status
